validasikan untuk edit seminar

+++ app/Http/Controllers/controller_seminar/EditSidangTesisController.php
@@ -36,7 +36,7 @@ class EditSidangTesisController extends Controller
     public function edit($id)
     {
         $seminar = ModelKompreS2::find(Crypt::decrypt($id));
-        $dosen = Dosen::select('id','encrypt_id', 'nama_dosen')->where('status', 'Aktif')->get();
+        $dosen = Dosen::select('id', 'encrypt_id', 'nama_dosen')->where('status', 'Aktif')->get();
         $lokasi = Lokasi::select('encrypt_id', 'nama_lokasi')->where('jenis_ruangan', 'Kelas')->get();
         $data = [
             'seminar' => $seminar,
@@ -46,7 +46,7 @@ class EditSidangTesisController extends Controller
         return view('koorS2.sidang.arsip.edit', $data);
     }
 
-    public function update(UpdateSidangTesisRequest $request, $id)
+    public function update(Request $request, $id)
     {
         try {
             $seminar = ModelKompreS2::find(Crypt::decrypt($id));
@@ -132,37 +132,41 @@ class EditSidangTesisController extends Controller
             $seminar->updated_at = date('Y-m-d H:i:s');
             $seminar->save();
 
-            $jadwal->tanggal = $request->tanggal;
-            $jadwal->jam_mulai = $request->jam_mulai;
-            $jadwal->jam_selesai = $request->jam_selesai;
-            $jadwal->id_lokasi = Crypt::decrypt($request->id_lokasi);
-            $jadwal->updated_at = date('Y-m-d H:i:s');
-            $jadwal->save();
+            if ($jadwal) {
+                $jadwal->tanggal = $request->tanggal;
+                $jadwal->jam_mulai = $request->jam_mulai;
+                $jadwal->jam_selesai = $request->jam_selesai;
+                $jadwal->id_lokasi = Crypt::decrypt($request->id_lokasi);
+                $jadwal->updated_at = date('Y-m-d H:i:s');
+                $jadwal->save();
+            }
 
-            $ba->no_ba = $request->no_ba;
-            $ba->nilai = $request->nilai;
-            $ba->nilai_mutu = $request->nilai_mutu;
-            $ba->ppt = $request->ppt;
-            if ($request->berkas_ba) {
-                if (file_exists('uploads/ba_sidang_tesis/' . $ba->berkas_ba)) {
-                    unlink('uploads/ba_sidang_tesis/' . $ba->berkas_ba);
+            if ($ba) {
+                $ba->no_ba = $request->no_ba;
+                $ba->nilai = $request->nilai;
+                $ba->nilai_mutu = $request->nilai_mutu;
+                $ba->ppt = $request->ppt;
+                if ($request->berkas_ba) {
+                    if (file_exists('uploads/ba_sidang_tesis/' . $ba->berkas_ba)) {
+                        unlink('uploads/ba_sidang_tesis/' . $ba->berkas_ba);
+                    }
+                    $file = $request->file('berkas_ba');
+                    $filename = $file->hashName();
+                    $file->move('uploads/ba_sidang_tesis', $filename);
+                    $ba->berkas_ba = $filename;
                 }
-                $file = $request->file('berkas_ba');
-                $filename = $file->hashName();
-                $file->move('uploads/ba_sidang_tesis', $filename);
-                $ba->berkas_ba = $filename;
-            }
-            if ($request->file_nilai) {
-                if (file_exists('uploads/nilai_sidang_tesis/' . $ba->file_nilai)) {
-                    unlink('uploads/nilai_sidang_tesis/' . $ba->file_nilai);
+                if ($request->file_nilai) {
+                    if (file_exists('uploads/nilai_sidang_tesis/' . $ba->file_nilai)) {
+                        unlink('uploads/nilai_sidang_tesis/' . $ba->file_nilai);
+                    }
+                    $file = $request->file('file_nilai');
+                    $filename = $file->hashName();
+                    $file->move('uploads/nilai_sidang_tesis', $filename);
+                    $ba->file_nilai = $filename;
                 }
-                $file = $request->file('file_nilai');
-                $filename = $file->hashName();
-                $file->move('uploads/nilai_sidang_tesis', $filename);
-                $ba->file_nilai = $filename;
+                $ba->updated_at = date('Y-m-d H:i:s');
+                $ba->save();
             }
-            $ba->updated_at = date('Y-m-d H:i:s');
-            $ba->save();
             DB::commit();
         } catch (\Throwable $th) {
             //throw $th;
